Respiration

+++ app/src/algos.h
@@ -15,7 +15,7 @@ void hpi_remove_close_peaks(int32_t *pn_locs, int32_t *pn_npks, int32_t *pn_x, i
 void hpi_sort_ascend(int32_t *pn_x, int32_t n_size);
 void hpi_sort_indices_descend(int32_t *pn_x, int32_t *pn_indx, int32_t n_size);
 
-void Resp_FilterProcess(int32_t * RESP_WorkingBuff, int32_t * CoeffBuf, int32_t* FilterOut);
-int32_t Resp_ProcessCurrSample(int32_t CurrAqsSample);
-void RESP_Algorithm_Interface(int32_t CurrSample,volatile uint32_t *RespirationRate);
-void Respiration_Rate_Detection(int32_t Resp_wave,volatile uint32_t *RespirationRate);
+void Resp_FilterProcess(int16_t * RESP_WorkingBuff, int16_t * CoeffBuf, int16_t* FilterOut);
+int16_t Resp_ProcessCurrSample(int16_t CurrAqsSample);
+void RESP_Algorithm_Interface(int16_t CurrSample,volatile uint8_t *RespirationRate);
+void Respiration_Rate_Detection(int16_t Resp_wave,volatile uint8_t *RespirationRate);
